+++ packages/gatsby/src/services/index.ts
@@ -13,4 +13,9 @@ export { writeOutRequires } from "./write-out-requires"
 export { calculateDirtyQueries } from "./calculate-dirty-queries"
 export { runStaticQueries } from "./run-static-queries"
 export { runPageQueries } from "./run-page-queries"
+
+export { waitUntilAllJobsComplete } from "../utils/wait-until-jobs-complete"
+export { runMutationBatch } from "./run-mutation-batch"
+export { recompile } from "./recompile"
+
 export * from "./types"