+++ packages/gatsby/src/services/source-nodes.ts
@@ -1,25 +1,29 @@
-import { IBuildContext } from "./"
 import sourceNodesAndRemoveStaleNodes from "../utils/source-nodes"
 import reporter from "gatsby-cli/lib/reporter"
-// import { findChangedPages } from "../utils/check-for-changed-pages"
-// import { IGatsbyPage } from "../redux/types"
+import { IDataLayerContext } from "../state-machines/data-layer/types"
+import { assertStore } from "../utils/assert-store"
+import { IGatsbyPage } from "../redux/types"
+import { findChangedPages } from "../utils/changed-pages"
 
 export async function sourceNodes({
   parentSpan,
   webhookBody,
   store,
-}: Partial<IBuildContext>): Promise<void> {
-  if (!store) {
-    reporter.panic(`No redux store`)
-  }
+  deferNodeMutation = false,
+}: Partial<IDataLayerContext>): Promise<{
+  deletedPages: string[]
+  changedPages: string[]
+}> {
+  assertStore(store)
+
   const activity = reporter.activityTimer(`source and transform nodes`, {
     parentSpan,
   })
   activity.start()
-  // const currentPages = new Map<string, IGatsbyPage>(store.getState().pages)
+  const currentPages = new Map<string, IGatsbyPage>(store.getState().pages)
   await sourceNodesAndRemoveStaleNodes({
     parentSpan: activity.span,
-    // deferNodeMutation: !!(webhookBody && Object.keys(webhookBody).length), // Coming soon
+    deferNodeMutation,
     webhookBody,
   })
 
@@ -31,31 +35,30 @@ export async function sourceNodes({
       .join(`, `)}]`
   )
 
-  // reporter.info(`Checking for deleted pages`)
+  reporter.verbose(`Checking for deleted pages`)
 
-  // Add this back when we enable page creation outside of onCreatePages
-  // const tim = reporter.activityTimer(`Checking for changed pages`)
-  // tim.start()
+  const tim = reporter.activityTimer(`Checking for changed pages`)
+  tim.start()
 
-  // const { changedPages, deletedPages } = findChangedPages(
-  //   currentPages,
-  //   store.getState().pages
-  // )
+  const { changedPages, deletedPages } = findChangedPages(
+    currentPages,
+    store.getState().pages
+  )
 
-  // reporter.info(
-  //   `Deleted ${deletedPages.length} page${deletedPages.length === 1 ? `` : `s`}`
-  // )
+  reporter.verbose(
+    `Deleted ${deletedPages.length} page${deletedPages.length === 1 ? `` : `s`}`
+  )
 
-  // reporter.info(
-  //   `Found ${changedPages.length} changed page${
-  //     changedPages.length === 1 ? `` : `s`
-  //   }`
-  // )
-  // tim.end()
+  reporter.verbose(
+    `Found ${changedPages.length} changed page${
+      changedPages.length === 1 ? `` : `s`
+    }`
+  )
+  tim.end()
 
   activity.end()
-  // return {
-  //   deletedPages,
-  //   changedPages,
-  // }
+  return {
+    deletedPages,
+    changedPages,
+  }
 }